requirement.txt
argcomplete==3.5.1
asgiref==3.8.1
beautifulsoup4==4.12.3
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
commitizen==3.29.1
crashtest==0.4.1
cryptography==43.0.3
decli==0.6.2
distlib==0.3.9
Django==4.2.16
django-active-link==0.2.2
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-restframework==0.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
keyring==24.3.1
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
nltk==3.9.1
packaging==24.1
pexpect==4.9.0
pip-tools==7.4.1
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.36
psycopg2==2.9.10
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.2.0
PyYAML==6.0.2
questionary==2.0.1
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
SecretStorage==3.3.3
setuptools==75.2.0
shellingham==1.5.4
soupsieve==2.6
sqlparse==0.5.1
termcolor==2.5.0
tomlkit==0.13.2
tqdm==4.66.5
trove-classifiers==2024.10.21.16
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
wheel==0.44.0
